Re-arrange the structure of modules
