+++ src/modules/health/get.controller.ts
@@ -1,6 +1,6 @@
 import type { Response } from 'express';
 
-import AbstractController from '../core/abstractController';
+import AbstractController from '../../core/abstractController';
 
 export class HealthController extends AbstractController {
     protected readonly requestSchema = null;
